+++ example/lib/theme.dart
@@ -1,6 +1,14 @@
 import 'package:fluent_ui/fluent_ui.dart';
+import 'package:system_theme/system_theme.dart';
 
 class AppTheme extends ChangeNotifier {
+  AccentColor _color = systemAccentColor;
+  AccentColor get color => _color;
+  set color(AccentColor color) {
+    _color = color;
+    notifyListeners();
+  }
+
   ThemeMode _mode = ThemeMode.system;
   ThemeMode get mode => _mode;
   set mode(ThemeMode mode) {
@@ -8,3 +16,13 @@ class AppTheme extends ChangeNotifier {
     notifyListeners();
   }
 }
+
+AccentColor get systemAccentColor => AccentColor('normal', {
+      'darkest': SystemTheme.accentInstance.darkest,
+      'darker': SystemTheme.accentInstance.darker,
+      'dark': SystemTheme.accentInstance.dark,
+      'normal': SystemTheme.accentInstance.accent,
+      'light': SystemTheme.accentInstance.light,
+      'lighter': SystemTheme.accentInstance.lighter,
+      'lightest': SystemTheme.accentInstance.lightest,
+    });